// src/utils/interfaces/questionType.ts
enum QuestionType {
    ShortAnswer = 'ShortAnswer',
    Paragraph = 'Paragraph',
    Dropdown = 'Dropdown',
    Checkbox = 'Checkbox',
    RadioButton = 'RadioButton',
    LinearScale = 'LinearScale',
    Description = 'Description',
    RadioButtonGrid = 'RadioButtonGrid',
}

export default QuestionType;